Split apart uber file
